Make waveform tests pass
Though the tests is working in Python-2, they failed in python-3 because of a
variable (x) not being defined ("test_waveform.py", line: 325). It's not clear
how it's defined when running Python-2, maybe by a side effect.

Looking at calls to the function, it is likely that the mysterious 'x' should
be 'x = args[0]'. Setting it so does make tests pass.